corrected transaction amount data type from integer to decimal, after flirting with float

+++ app/db/budgie_budget.sql
@@ -14,7 +14,7 @@ CREATE TABLE merchants (
 
 CREATE TABLE transactions (
     id SERIAL PRIMARY KEY,
-    amount INT,
+    amount DECIMAL,
     merchant_id INT REFERENCES merchants(id),
     tag_id INT REFERENCES tags(id)
 );
